Final Edits During Competition
This was the final edits made to the code during the competition. If I
remember correctly, the issues we were having were the wheels were too
small and they couldn't get over the electrical tape on the track.
Another issue was the claw would sometimes not drop the fish into the
bin correctly. Also the conveyor would sometimes skip over a bin because
the IR sensor would miss it (the program was running to slow I think).
We also didn't finish the robot to be able to go around the whole track,
because we wanted to make sure to score at least SOME points. Overall,
it was a success! If we had one more week, I think our robot would be
able to make it fully around the track and pick up all the fish. I am
not sure if we could complete it in the required amount of time.